Fix the invalid JSON for some missions
For some Disney missions we had an invalid JSON because last charater
was removed - the "}". With this hack it works. Also since the 'a'
option means that the string is null-padded to the given length,
the previously working missions would work too. Just the broken ones
would start working (at least that's what I saw in the test in sim)